Split up database roundtrip tests into individual files
